+++ resources/views/auth/passwords/email.blade.php
@@ -28,12 +28,10 @@
                 @endif
             </div>
         </div>
-        <div class="row">
-            <div class="form-group col-md-6 col-xs-12 offset-md-6">
-                <button class="btn btn-default btn-block" type="submit">
-                    <i class="icon-envelope text-xs m-r-xs"></i>  {{ __('Send Password Reset Link') }}
-                </button>
-            </div>
+        <div class="form-group">
+            <button class="btn btn-default btn-block" type="submit">
+                <i class="icon-envelope text-xs m-r-xs"></i>  {{ __('Send Password Reset Link') }}
+            </button>
         </div>
     </form>
 @endsection