+++ resources/views/layouts/app.blade.php
@@ -11,7 +11,7 @@
 
     @stack('head')
 
-    <meta name="turbolinks-cache-control" content="no-preview">
+    <meta name="turbolinks-cache-control" content="no-cache">
     <meta name="turbolinks-root" content="{{Dashboard::prefix()}}">
     <meta name="dashboard-prefix" content="{{Dashboard::prefix()}}">