+++ src/Platform/Dashboard.php
@@ -11,7 +11,7 @@ class Dashboard
     /**
      * ORCHID Version.
      */
-    public const VERSION = '3.10.1';
+    public const VERSION = '3.10.3';
 
     /**
      * The Dashboard configuration options.